requirements.txt
kafka-python==2.0.2
oauthlib==1.0.3
requests==2.25.1
six==1.10.0
tweepy==3.10.0
configparser==3.3.0.post2
spacy==3.1.3
spacy-legacy==3.0.8
en-core-web-sm==3.1.0
numpy==1.21.2
packaging==21.0
pandas==1.3.3
nltk==3.6.2